refactor: split up local cmd.go file
